move SourcePath, colocate AllowListItem with their implementation
