+++ src/descriptor.rs
@@ -1,7 +1,7 @@
 mod data_package;
 mod table_schema;
 
-pub use data_package::{DataPackage, DataResource, Name, SourcePath, TableSource};
+pub use data_package::{DataPackage, DataResource, Name, TableSource};
 pub use table_schema::{
     AnyFieldType, ArrayFieldType, BooleanFieldType, Constraints, DateFieldType, DateTimeFieldType,
     GeoPointFieldType, NumberFieldType, ObjectFieldType, StringFieldFormat, StringFieldType,
